+++ backend/.env.example
@@ -2,6 +2,8 @@
 DATABASE_URL="postgres://<user>:<pass>@localhost:5432/postgres?schema=public"
 COIN_GECKO_URL=""
 
-ACCESS_TOKEN_SECRET=""
+ACCESS_TOKEN_EXPIRATION=""
+REFRESH_TOKEN_EXPIRATION=""
 
+ACCESS_TOKEN_SECRET=""
 ACCESS_TOKEN_PUBLIC=""
