--- backend/src/routes/auth.route.js
@@ -1,23 +1,0 @@
-const express = require("express");
-const {
-    register,
-    login,
-    refreshAccessToken,
-    logout,
-} = require("../controllers/auth.controller");
-
-const router = express.Router();
-
-/* Register user route */
-router.post("/register", register);
-
-/* Login user route */
-router.post("/login", login);
-
-/* Refresh access token route */
-router.get("/refresh", refreshAccessToken);
-
-/* Logout user route */
-router.delete("/logout", logout);
-
-module.exports = router;
